face_key/requirements.txt
Django~=3.0.7
django-allauth~=0.42.0